Sonarqube.

+++ src/main/java/com/chadfield/shogiexplorer/utils/MathUtils.java
@@ -3,10 +3,10 @@ package com.chadfield.shogiexplorer.utils;
 
 
 public class MathUtils {
-    public static final long KOMA_X = 43;
-    public static final long KOMA_Y = 48;
-    public static final long BOARD_XY = 9;
-    public static final long COORD_XY = 20;
+    public static final int KOMA_X = 43;
+    public static final int KOMA_Y = 48;
+    public static final int BOARD_XY = 9;
+    public static final int COORD_XY = 20;
     
     private MathUtils() {
         throw new IllegalStateException("Utility class");
